Stop KCC-text logo from displaying on mobile on landing pages...
the 'Kankakee Community College' text logo takes up to much space on the global nav of landing pages (nav-landing.html).
It was causing the phone No. to be pushed outside of the nav element.
Previously, this change was made to nav-global.html...This commit restores nav-global.html and instead adds the fix to nav-ladning.html
nav-global.html has enough space to afford the KCC logo KCC-text logo & search icon.