Fix issue w/ webpack bundle finishing after gulp...
In running gulp & webpack in parallel webpack is finishing after gulp.
This means gulp doesn't get a chance to copy the JS bundle into the _site/ dir (which results in broken JS links)

The solution: Make the gulp command a sequence of 2 commands.
It runs `$ sleep 1` to give webpack a chance to bundle, then `$ gulp`.

This change was made to `$ npm run production` & `$ npm run dev`.