Add a class with forloop.length (the \# of slides) to fix styling on longer sliders.
For slide headings to display correctly, the `.hero-slider__slider--heading-container` class's `width:` needs to adjusted based on how many slides there are

e.g. .hero-slider__slider--heading-container { width: 10%; } is fine for a slider with 4 slides, but is completely off for a slider with 8 slides (needs a width more like 5%).

Because this code has to be reuseable (it's a reuseable `_includes` file), we need a way to adjust that width depending on the # of slides.
Liquid's {{ forloop.length }} can be used to get the slide \# as a digit & then classes can be written to adjust for the \# of slides.