Restore webpack HashPlugin...
IN ORDER FOR THE JEKYLL/WEBPACK TIMING IN OUR BUILD PROCESS TO WORK WE NEED THE HASHPLUGIN:

The Webpack HashPlugin is not used in this project like it is in the 'kcc-startup-template'.
In the 'kcc-startup-template', the generated hash is appended to the stylesheet <link>'s for cache-busting purposes.

HOWEVER, the other function of this generated hash is to kick off-jekyll ==> Jekyll watches the `_data` dir, by default (the same dir we are injecting the hash file into).
Everytime webpack's bundle changes (i.e. changes to the JS were made), the hash is updated. Jekyll sees this, & handles the updating of `_site` build (consequently, updating the changed JS).