Create Custom JS to collapse the mobile menu when resizing the window...
This is to prevent the mobile menu from staying open when it reaches screen sizes where the menu is never in
a collapsed state but always expanded: (BS$ large devices: 992px and above).

Create custom JS file to collapse the mobile menu:
1. Use eventListener to watch for window resizes.
2. If the window.innerWidth is greater than Bootstrap 4 Large devices (992px when it should NOT be a mobile
menu anymore.) proceed to the next function.
3. If the menu's collapse div has the 'show' class it will fire a function to collapse the menu using BS4
`.collapse()`:
```javascript
$(<COLLAPSABLE_EL>).collapse('hide');
```