Filter out clicks on dropdown menu items...
closeMenuOnClick.js closes the menu for dropdown menu items too. Instead, it should stay open if any item with
the BS4 .dropdown-toggle class is clicked. This will allow mobile users the same convenience in using the
desktop "mega menu".

Add an if statement to bail-out of the `closeMenuOnClick` functions, if the clicked target is a
dropdown-toggle:
```javascript
if ( e.target.classList.contains('dropdown-toggle') )
  return;
```