Replace footer url with `http://` protocol for `www.kcc.edu`...
To avoid a broken link while visiting the sites on KCC computers,
the `https://www.kcc.edu` should be `http://www.kcc.edu`.

The URL and link-text in footer is pulled from the `public-url` entry in `_config.yml`.
A Liquid replace filter removes the `https://` protocol to generate the link text.
An additional replace filter was added to remove `http://` protocols as well.