Fix mega-nav bugginess when resizing window with dropdown item open...
If a dropdown item is clicked to the open state,
and the screen is shrunk-down to the point it has the mobile version of the menu,
the mobile menu has to get toggled open to show the same dripdown items that were activated.
Otherwise, you get a dropwdown menu stuck open inside the collapsed menu.