Switch accordion +/- indicator to CSS-only solution.
The kcc-gem-theme's accordion component includes blue-circle-icons with
'+' and '-' to indicate if the accordion is open or closed.

The accordion component had used JavaScript to toggle the icons from a '+' to a '-' (and vice-versa).

To improve performance & simplify/reduce code, this functionality was switched to a CSS-only solution.
By leveraging pseudo-elements plus using Bootstrap's `.collapsed` class, the same indicator-icon-effect can be attained.