Add custom JS to underline the current site in the navigation...
Create a JS module to underline the current site in the navigation
i.e. If user is anywhere in /academics site, "Academics" link is underlined in the nav

This feature addition is to provide the user with a visual reference as to where they are in the site
Breadcrumbs without the breadcurmbs.