Add lazy loading of slick and DataTables JS modules...
Use dynamic imports to lazily load JS modules, if the page call for it.

Add `chunkFilename: '[name].bundle.js'` and `publicPath: 'assets/js/theme/dist/'` to webpacks output config.
This allows chunks to be broken out into their own files and imported when needed.

Add dynamic imports with webpack "magic comments.":
Wrap imports with if-statement that checks if any elements in the page match the querySelectorString for hero-slider or DataTable markup.
If markup exsists for those components in the current page, the appropriate module gets loaded.
If it doesn't exist, it won't loaded an ENTIRE MODULE'S-WORTH OF JS.