Sort each list of links in the mega-menu alphabetically via Liquid...
To sort the dropdown items alphabetically, I had to get rid of the unused "dropdown-divider" option.
Having a divider to separate links into sections requieres the links stay in the written order.

Use `{% assign alphabetical_links = item.dropdown | sort: "text" %}` to sort by the text key, which is the link-text.
Strings are sorted alphabetically in Liqyuid.