Add browser history entries when navigating through accordions/tabs...
Some pages have a lot of accordion items where it is easy to lose track of
which items have already been visitied.

This commit adds a feature to track navigation when clicking through tabs
or accordion cards. This allows the user to navigate backwards through the
accordions/tabs they've visited--using the browser's back button.

This commit also alters the user's URL when clicking through tabs and accordions.
This has the added benefit of providing any user with a link to specific content
(within  an accordion or tab)

- Add window.history entries when navigating through accordions or tabs:
  - Use `history.pushState()` to add history entries for each accordion
    or tab click.
- Create simple module (`checkForAccordionOrTab`) that checks for tab or accordion
  in the current page and then lazy loads the `contentHashLink` module if appropriate.
  - Remove all references and imports of `contentHashLink.js`. Replace with new
    `checkForAccordionOrTab.js` module.