Add `alternate_id` feature to accordion...
This adds the ability to force an accordion card to use a specific
string as the ID, supplied by the developer, which is built into the HTML.

Normally, the ID is automatically generated from the `card_heading` key.
Sometimes, this yields unsatisfactory results especially since the ID may
appear in URLs.

- Add an optional new `alternate_id` key to the accordion component:
  - If this key is present it used used as the ID for the accordion's
    button. E.g. `alternate_id: must-be-a-unique-valid-id`